spinner was added
