+++ src/components/Main/Main.scss
@@ -1,0 +1,6 @@
+.Main {
+  display: flex;
+  justify-content: center;
+  flex-grow: 1;
+  width: 100%;
+}
