+++ src/components/common/BackgroundAnimation/BackgroundAnimation.scss
@@ -1,4 +1,4 @@
-@import "./src/assets/variables";
+@import "../../../assets/variables";
 
 .Hero {
   position: absolute;
