+++ src/components/common/BackgroundAnimation/BackgroundAnimation.tsx
